start on mst to eulerian
